+++ R/sits_indexes.R
@@ -12,12 +12,15 @@
 #' @export
 sits_savi <- function(data) {
 
+    # set caller to show in errors
+    .check_set_caller("sits_savi")
+
     bands <- sits_bands(data)
     bands_savi <- c("NIR", "RED")
     .check_chr_within(
         x = bands_savi,
         within = bands,
-        msg = "sits_savi: not enough bands to compute"
+        msg = "not enough bands to compute"
     )
 
     data <- sits_mutate_bands(data,
@@ -42,13 +45,16 @@ sits_savi <- function(data) {
 #' @export
 sits_ndwi <- function(data) {
 
+    # set caller to show in errors
+    .check_set_caller("sits_ndwi")
+
     bands <- sits_bands(data)
     bands_ndwi <- c("NIR", "MIR")
 
     .check_chr_within(
         x = bands_ndwi,
         within = bands,
-        msg = "sits_ndwi: not enough bands to compute."
+        msg = "not enough bands to compute."
     )
 
     data <- sits_mutate_bands(data, NDWI = (1.5) * (NIR - MIR) / (NIR + MIR))