+++ R/sits_labels_set.R
@@ -30,6 +30,9 @@
 #'
 `sits_labels<-` <- function(data, value) {
 
+    # set caller to show in errors
+    .check_set_caller("sits_labels")
+
     # get the meta-type (sits or cube)
     data <- .config_data_meta_type(data)
 
@@ -47,26 +50,26 @@
 
     # check if value is an atomic vector
     .check_chr_type(x = value,
-                    msg = "sits_labels: value must be a character vetor")
+                    msg = "value must be a character vetor")
 
     # check if length is correct
     .check_length(
         x = labels,
         len_max = length(value),
         len_min = length(value),
-        msg = "sits_labels: informed labels have a different expected length"
+        msg = "informed labels have a different expected length"
     )
 
     # check if there are no NA
     .check_that(
         x = all(!is.na(value)),
-        msg = "sits_labels: invalid labels value"
+        msg = "invalid labels value"
     )
 
     # check if there are empty strings
     .check_that(
         x = any(trimws(value) != ""),
-        msg = "sits_labels: invalid labels value"
+        msg = "invalid labels value"
     )
 
     names(value) <- labels