+++ R/sits_source_api_bdc.R
@@ -44,6 +44,9 @@
                                        stac_query,
                                        tiles = NULL) {
 
+    # set caller to show in errors
+    .check_set_caller(".source_items_new.bdc_cube")
+
     # if specified, a filter per tile is added to the query
     if (!is.null(tiles))
         stac_query <- rstac::ext_query(q = stac_query, "bdc:tile" %in% tiles)
@@ -54,7 +57,7 @@
     # check if matched items
     .check_that(
         x = rstac::items_matched(items_info) > 0,
-        msg = ".source_items_new.bdc_cube: no items matched the query criteria."
+        msg = "no items matched the query criteria."
     )
 
     # if more than 2 times items pagination are found the progress bar
@@ -74,24 +77,6 @@
                                                collection = NULL) {
 
     rstac::items_group(items, field = c("properties", "bdc:tiles"))
-}
-
-#' @keywords internal
-#' @export
-.source_items_get_sensor.bdc_cube <- function(source,
-                                              items, ...,
-                                              collection = NULL) {
-
-    items[["features"]][[1]][[c("properties", "instruments")]]
-}
-
-#' @keywords internal
-#' @export
-.source_items_get_satellite.bdc_cube <- function(source,
-                                                 items, ...,
-                                                 collection = NULL) {
-
-    items[["features"]][[1]][[c("properties", "platform")]]
 }
 
 #' @keywords internal