+++ R/sits_source_api_stac.R
@@ -54,6 +54,9 @@
                                    start_date,
                                    end_date) {
 
+    # set caller to show in errors
+    .check_set_caller(".source_cube.stac_cube")
+
     items_query <- .stac_items_query(source = source,
                                      collection = collection,
                                      name = name,
@@ -117,6 +120,9 @@
                                              items, ...,
                                              collection = NULL) {
 
+    # set caller to show in errors
+    .check_set_caller(".source_items_fileinfo.stac_cube")
+
     file_info <- purrr::map_dfr(items$features, function(item){
 
         date <- suppressWarnings(
@@ -139,22 +145,22 @@
 
         .check_that(
             x = !is.na(date),
-            msg = ".source_cube: invalid date format."
+            msg = "invalid date format."
         )
 
         .check_that(
             x = is.character(bands),
-            msg = ".source_cube: invalid band format."
+            msg = "invalid band format."
         )
 
         .check_that(
             x = is.numeric(res),
-            msg = ".source_cube: invalid res format."
+            msg = "invalid res format."
         )
 
         .check_that(
             x = is.character(paths),
-            msg = ".source_cube: invalid path format."
+            msg = "invalid path format."
         )
 
         tidyr::unnest(
@@ -178,6 +184,8 @@
                                          items,
                                          file_info, ...) {
 
+    # set caller to show in errors
+    .check_set_caller(".source_items_cube.stac_cube")
 
     t_bbox <- .source_items_tile_get_bbox(source = source,
                                           tile_items = items, ...,
@@ -191,7 +199,7 @@
     )
 
     .check_num_type(x = t_bbox,
-                    msg = ".source_items_cube.stac_cube: bbox must be numeric.")
+                    msg = "bbox must be numeric.")
 
     t_size <- .source_items_tile_get_size(source = source,
                                           tile_items = items, ...,
@@ -199,13 +207,12 @@
     .check_chr_within(
         x = names(t_size),
         within = c("nrows", "ncols"),
-        msg = paste(".source_items_cube.stac_cube: size must be have",
-                    "'nrows' and 'ncols' names.")
+        msg = "size must be have 'nrows' and 'ncols' names."
     )
 
     .check_num_type(
         t_size,
-        msg = ".source_items_cube.stac_cube: size must be numeric."
+        msg = "size must be numeric."
     )
 
     # tile name
@@ -215,8 +222,7 @@
 
     .check_chr_type(
         x = t_name,
-        msg = paste(".source_items_cube.stac_cube: name must be a",
-                    "character value.")
+        msg = "name must be a character value."
     )
 
     t_crs <- .source_items_tile_get_crs(source = source,
@@ -224,36 +230,15 @@
                                         collection = collection)
     .check_that(
         x = is.character(t_crs) || is.numeric(t_crs),
-        msg = paste(".source_items_cube.stac_cube: name must be a",
-                    "character or numeric value.")
-    )
-
-    t_sat <- .source_items_get_satellite(source = source,
-                                         items = items, ...,
-                                         collection = collection)
-
-    .check_chr_type(
-        x = t_sat,
-        msg = paste(".source_items_cube.stac_cube: satellite name must be",
-                    "a character value.")
-    )
-
-    t_sensor <- .source_items_get_sensor(source = source,
-                                         items = items, ...,
-                                         collection = collection)
-
-    .check_chr_type(
-        x = t_sensor,
-        msg = paste(".source_items_cube.stac_cube: sensor name must be a",
-                    "character value.")
+        msg = "name must be a character or numeric value."
     )
 
     tile <- .sits_cube_create(
         name       = name[[1]],
         source     = source[[1]],
         collection = collection[[1]],
-        satellite  = t_sat[[1]],
-        sensor     = t_sensor[[1]],
+        satellite  = .source_collection_satellite(source, collection),
+        sensor     = .source_collection_sensor(source, collection),
         tile       = t_name[[1]],
         bands      = unique(file_info[["band"]]),
         nrows      = t_size[["nrows"]],